+++ test/Spec.hs
@@ -6,6 +6,7 @@ import Test.HUnit
 
 import qualified Spec.AndReq as AndReq
 import qualified Spec.Constructors as Constructors
+import qualified Spec.MonadPlus as MonadPlus
 import qualified Spec.NotReq as NotReq
 import qualified Spec.OrReq as OrReq
 import qualified Spec.ReqIf as ReqIf
@@ -25,4 +26,5 @@ tests = TestList
   , ReqIf.tests
   , ReqLift.tests
   , ReqLift2.tests
+  , MonadPlus.tests
   ]